% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd}
\alias{rd}
\title{Read a Tab-Separated Value (TSV) file into a data frame.}
\usage{
rd(f, ...)
}
\arguments{
\item{f}{The path to the TSV file.}

\item{...}{Additional arguments to pass to \code{read.table}.}
}
\value{
A data frame containing the data from the TSV file.
}
\description{
This is a wrapper around \code{read.table} specifically configured for TSV files,
with headers, and treating strings as characters rather than factors.
}
\seealso{
\code{\link[=read.table]{read.table()}}
}
\keyword{internal}
