% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pros.R
\name{read_pros}
\alias{read_pros}
\title{Read process definitions for a model based on the number of phytoplankton groups.}
\usage{
read_pros(file, phyto_groups = 3)
}
\arguments{
\item{file}{The path to the TSV file containing process definitions.}

\item{phyto_groups}{An integer, either 2 or 3, specifying the number of
phytoplankton groups in the model version.}
}
\value{
A data frame with process definitions, with column names adjusted
according to \code{phyto_groups}.
}
\description{
This function reads a TSV file containing process definitions for a model.
It handles different column names based on whether the model uses 2 or 3
phytoplankton groups, specifically by renaming columns for the 3-group case.
}
\keyword{internal}
