% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_replacements.R
\name{read_replacements}
\alias{read_replacements}
\title{Read replacement rules for a model based on the number of phytoplankton groups.}
\usage{
read_replacements(file, phyto_groups = 3)
}
\arguments{
\item{file}{The path to the TSV file containing replacement rules.}

\item{phyto_groups}{An integer, either 2 or 3, specifying the number of
phytoplankton groups in the model version.}
}
\value{
A data frame with replacement rules, with column names adjusted
according to \code{phyto_groups}.
}
\description{
This function reads a TSV file containing replacement rules for a model.
It handles different column names based on whether the model uses 2 or 3
phytoplankton groups, specifically by renaming columns for the 3-group case.
}
\seealso{
\code{\link[=rd]{rd()}}
}
\keyword{internal}
