% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmorodeo-package.R
\docType{package}
\name{salmoRodeo-package}
\alias{salmoRodeo}
\alias{salmoRodeo-package}
\title{A 1-layer ecological lake model in Petersen-Gujer matrix notation}
\description{
The package implements a 1-layer version of the ecological lake
model SALMO using a tabular Petersen-Gujer matrix description. It simulates
main components of the pelagic foodweb in a shallow, non-stratified lake.
}
\details{
Usage of the package requires installed build tools,
Gnu C/C++ and Gnu Fortran on Linux or
\href{https://cran.r-project.org/bin/windows/Rtools/}{RTools} on Windows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tpetzoldt.github.io/salmoRodeo/}
}

}
\author{
Justus Möller

Thomas Petzoldt  \email{thomas.petzoldt@tu-dresden.de}
}
\keyword{internal}
