% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pars.R
\name{set_pars}
\alias{set_pars}
\title{Set or change parameter values in a model object.}
\usage{
set_pars(model, new_pars = NULL, default_pars = model$getPars())
}
\arguments{
\item{model}{A model object that has \code{getPars()} and \code{setPars()} methods
(e.g., a \code{rodeo} model object).}

\item{new_pars}{A named vector of new parameter values to set. Its names
should correspond to parameters in the model. If \code{NULL}, no parameters
are changed beyond \code{default_pars}.}

\item{default_pars}{An optional named vector representing the default
parameters. If not provided, \code{model$getPars()} will be used as the base
onto which \code{new_pars} are applied.}
}
\value{
The modified model object with updated parameters.
}
\description{
This function updates the parameters of a model object. It accepts a named
vector of new parameters and uses the \code{replace_values} helper function to
merge these into the existing or default parameters, preserving other values
and their order.
}
\seealso{
\code{\link[=replace_values]{replace_values()}} \code{\link[=set_pars2]{set_pars2()}}
}
