% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pars2.R
\name{set_pars2}
\alias{set_pars2}
\title{Set or change parameter values in a model object.}
\usage{
set_pars2(model, default_pars = model$getPars(), ...)
}
\arguments{
\item{model}{A model object that has \code{getPars()} and \code{setPars()} methods.}

\item{default_pars}{An optional named vector representing the default
parameters. If not provided, \code{model$getPars()} will be used.}

\item{...}{Named arguments representing the new parameter values to set.
These will override existing values in \code{default_pars}.}
}
\value{
The modified model object with updated parameters.
}
\description{
This function updates the parameters of a model object. You can provide
new parameter values directly as named arguments via the ellipsis (\code{...}).
}
\details{
This function works similar to \code{set_pars}, but \code{name = value}-pairs
can be given directly, without an enclosing combine \code{c()}.
}
\seealso{
\code{\link[=replace_values]{replace_values()}} \code{\link[=set_pars]{set_pars()}}
}
